Rename API classes
